feat(lit-ui-router): per-element registration modules
ui-view.register.ts / ui-router.register.ts each define one element and
carry its own tag-map entry, restoring the single-element registration
the decorators used to give spec graphs. register.ts stays as a thin
aggregate: it is the public /register subpath, so the barrel and the
register-everything consumers share one module (index -> register ->
per-element). Both per-element modules get exports subpaths and join
the sideEffects list; specs import exactly the element they mount, and
register.spec.ts pins the granularity (ui-view defined, ui-router not).
