+++ docs/api/index.md
@@ -17,6 +17,7 @@ pnpm add lit-ui-router
 | `import { ... } from 'lit-ui-router'`      | Full API. Any value import registers the `<ui-router>`/`<ui-view>` custom elements as a side effect.        |
 | `import { ... } from 'lit-ui-router/pure'` | The same full API — element classes included — with no registration and no `HTMLElementTagNameMap` globals. |
 | `import 'lit-ui-router/register'`          | Registration only: defines `<ui-router>`/`<ui-view>` and carries their `HTMLElementTagNameMap` entries.     |
+| `import 'lit-ui-router/ui-view.register'`  | Single-element registration: defines just that element with its tag-map entry (`ui-router.register` ditto). |
 | `import type { ... } from 'lit-ui-router'` | Types are erased at compile time — always free, from any entry.                                             |
 
 ## Quick Start
