+++ packages/lit-ui-router/src/specs/transition-controller.spec.ts
@@ -7,7 +7,7 @@ import {
   TransitionController,
   TransitionControllerOptions,
 } from '../transition-controller.js';
-import '../register.js';
+import '../ui-router.register.js';
 import { UIRouterLit } from '../core.js';
 import { LitStateDeclaration } from '../interface.js';
 import {
