+++ packages/lit-ui-router/src/specs/ui-sref-active.spec.ts
@@ -13,6 +13,7 @@ import {
 } from '../ui-sref-active.js';
 import { uiSref, UI_SREF_TARGET_EVENT } from '../ui-sref.js';
 import { UIRouterLitElement } from '../ui-router.js';
+import '../ui-view.register.js';
 import { UIRouterLit } from '../core.js';
 import { LitStateDeclaration } from '../interface.js';
 import {
